# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Atom CMS browser API",
    "version": "1.0.0",
    "description": "Versioned browser API available in full and headless modes. All authentication operations use the existing cookie session. Requests must send Accept: application/json; unsafe requests require X-XSRF-TOKEN. Auth endpoints keep their native Fortify/Atom response shape. Public v1 data is allowlisted; monetary response amounts are integer minor units. Dynamic hotel validation rules and emulator capability gates remain server authoritative."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "operationId": "searchUsers",
        "summary": "searchUsers",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "username": {
                            "type": "string"
                          },
                          "look": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "username",
                          "look"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Request rejected. Inspect code when present; validation details are keyed by field.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/me/referral-claim": {
      "post": {
        "operationId": "claimReferral",
        "summary": "claimReferral",
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "default": {
            "description": "Request rejected. Inspect code when present; validation details are keyed by field.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/badges": {
      "get": {
        "operationId": "getBadgePrice",
        "summary": "getBadgePrice",
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "cost": {
                          "type": "integer"
                        },
                        "currency": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "cost",
                        "currency"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Request rejected. Inspect code when present; validation details are keyed by field.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "buyBadge",
        "summary": "buyBadge",
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "badge_url": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "badge_url"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Request rejected. Inspect code when present; validation details are keyed by field.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "badge_data is a validated GIF data URL; badge dimensions/palette follow the shared badge purchase rules.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "badge_data": {
                    "type": "string"
                  },
                  "badge_name": {
                    "type": "string"
                  },
                  "badge_description": {
                    "type": "string"
                  }
                },
                "required": [
                  "badge_data",
                  "badge_name",
                  "badge_description"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "laravel_session",
        "description": "Cookie name is configurable. Obtain the cookie through Sanctum CSRF initialization and Fortify login; this API does not accept bearer tokens."
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "vote_url": {
            "type": "string",
            "description": "Present only for vote_required; backend-generated vote destination."
          }
        },
        "required": [
          "code",
          "message"
        ],
        "additionalProperties": false
      },
...
    }
  }
}